Improvements post-review + implementing zero-arg .save()
