+++ .gitignore
@@ -37,3 +37,4 @@ pip-wheel-metadata/
 /qpdf-7.0.0/
 /tests/resources/private/
 /tmp_rpms/
+*.dll
