+++ src/qpdf/qpdf.cpp
@@ -657,104 +657,7 @@ void init_qpdf(py::module &m)
 
             )~~~"
         )
-        .def("save",
-            save_pdf,
-            R"~~~(
-            Save all modifications to this :class:`pikepdf.Pdf`.
-
-            Args:
-                filename (str or stream): Where to write the output. If a file
-                    exists in this location it will be overwritten. The file
-                    should not be the same as the input file, because data from
-                    the input file may be lazily loaded; as such overwriting
-                    in place will null-out objects.
-
-                static_id (bool): Indicates that the ``/ID`` metadata, normally
-                    calculated as a hash of certain PDF contents and metadata
-                    including the current time, should instead be generated
-                    deterministically. Normally for debugging.
-                preserve_pdfa (bool): Ensures that the file is generated in a
-                    manner compliant with PDF/A and other stricter variants.
-                    This should be True, the default, in most cases.
-
-                min_version (str or tuple): Sets the minimum version of PDF
-                    specification that should be required. If left alone QPDF
-                    will decide. If a tuple, the second element is an integer, the
-                    extension level. If the version number is not a valid format,
-                    QPDF will decide what to do.
-                force_version (str or tuple): Override the version recommend by QPDF,
-                    potentially creating an invalid file that does not display
-                    in old versions. See QPDF manual for details. If a tuple, the
-                    second element is an integer, the extension level.
-                fix_metadata_version (bool): If ``True`` (default) and the XMP metadata
-                    contains the optional PDF version field, ensure the version in
-                    metadata is correct. If the XMP metadata does not contain a PDF
-                    version field, none will be added. To ensure that the field is
-                    added, edit the metadata and insert a placeholder value in
-                    ``pdf:PDFVersion``. If XMP metadata does not exist, it will
-                    not be created regardless of the value of this argument.
-
-                object_stream_mode (pikepdf.ObjectStreamMode):
-                    ``disable`` prevents the use of object streams.
-                    ``preserve`` keeps object streams from the input file.
-                    ``generate`` uses object streams wherever possible,
-                    creating the smallest files but requiring PDF 1.5+.
-
-                compress_streams (bool): Enables or disables the compression of
-                    stream objects in the PDF. Metadata is never compressed.
-                    By default this is set to ``True``, and should be except
-                    for debugging.
-
-                stream_decode_level (pikepdf.StreamDecodeLevel): Specifies how
-                    to encode stream objects. See documentation for
-                    ``StreamDecodeLevel``.
-
-                normalize_content (bool): Enables parsing and reformatting the
-                    content stream within PDFs. This may debugging PDFs easier.
-
-                linearize (bool): Enables creating linear or "fast web view",
-                    where the file's contents are organized sequentially so that
-                    a viewer can begin rendering before it has the whole file.
-                    As a drawback, it tends to make files larger.
-
-                qdf (bool): Save output QDF mode.  QDF mode is a special output
-                    mode in QPDF to allow editing of PDFs in a text editor. Use
-                    the program ``fix-qdf`` to fix convert back to a standard
-                    PDF.
-
-                progress (callable): Specify a callback function that is called
-                    as the PDF is written. The function will be called with an
-                    integer between 0-100 as the sole parameter, the progress
-                    percentage. This function may not access or modify the PDF
-                    while it is being written, or data corruption will almost
-                    certainly occur.
-
-                encryption (pikepdf.models.Encryption or bool): If ``False``
-                    or omitted, existing encryption will be removed. If ``True``
-                    encryption settings are copied from the originating PDF.
-                    Alternately, an ``Encryption`` object may be provided that
-                    sets the parameters for new encryption.
-
-            You may call ``.save()`` multiple times with different parameters
-            to generate different versions of a file, and you *may* continue
-            to modify the file after saving it. ``.save()`` does not modify
-            the ``Pdf`` object in memory, except possibly by updating the XMP
-            metadata version with ``fix_metadata_version``.
-
-            .. note::
-
-                :meth:`pikepdf.Pdf.remove_unreferenced_resources` before saving
-                may eliminate unnecessary resources from the output file, so
-                calling this method before saving is recommended. This is not
-                done automatically because ``.save()`` is intended to be
-                idempotent.
-
-            .. note::
-
-                pikepdf can read PDFs will incremental updates, but always
-                any coalesces incremental updates into a single non-incremental
-                PDF file when saving.
-            )~~~",
+        .def("_save", save_pdf, "",
             py::arg("filename"),
             py::arg("static_id")=false,
             py::arg("preserve_pdfa")=true,
